feat : dealHistory save, find 기능을 테스트 완료했습니다.

+++ src/main/kotlin/krescere/jipsayobackend/dto/common/KakaoAddressResponse.kt
@@ -1,11 +1,13 @@
 package krescere.jipsayobackend.dto.common
 
+import java.math.BigDecimal
+
 class KakaoAddressResponse(
     val jibunAddress: String,
     val roadAddress: String,
     val hangCode: Long,
     val danjiName: String,
     val postCode: Int,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: BigDecimal,
+    val longitude: BigDecimal,
 )
